split into 2 components, styled historical view
