+++ backend/data-service/app/api/routes.py
@@ -3,11 +3,16 @@ from fastapi import APIRouter, Depends, Query
 from sqlalchemy.orm import Session
 
 from app.db.database import get_db
-from app.db.repositories.resale import get_all_resale_records
+from app.db.repositories.resale import (
+    get_all_resale_records,
+    get_all_towns,
+    get_all_flat_types,
+)
 from app.schemas.schemas import ResaleRecordResponse
 
 
 router = APIRouter()
+opts_router = APIRouter()
 
 
 @router.get("/history", response_model=List[ResaleRecordResponse])
@@ -21,3 +26,13 @@ def get_resale_data(
     return get_all_resale_records(
         db, limit=limit, offset=offset, flat_type=flat_type, town=town
     )
+
+
+@opts_router.get("/town", response_model=List[str])
+def get_available_towns(db: Session = Depends(get_db)):
+    return get_all_towns(db)
+
+
+@opts_router.get("/flat_type", response_model=List[str])
+def get_available_flat_types(db: Session = Depends(get_db)):
+    return get_all_flat_types(db)
