+++ docker-compose.yaml
@@ -45,15 +45,15 @@ services:
     ports:
       - "8000:8000"
 
-  frontend:
-    build:
-      context: ./frontend
-    ports:
-      - "5173:5173"
-    volumes:
-      - ./frontend:/app:rw,z
-    depends_on:
-      - data-service
-      - model-service
+  # frontend:
+  #   build:
+  #     context: ./frontend
+  #   ports:
+  #     - "5173:5173"
+  #   volumes:
+  #     - ./frontend:/app:rw,z
+  #   depends_on:
+  #     - data-service
+  #     - model-service
 volumes:
   pgdata:
