+++ frontend/package.json
@@ -13,6 +13,7 @@
     "@tailwindcss/vite": "^4.1.8",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-icons": "^5.5.0",
     "tailwindcss": "^4.1.8"
   },
   "devDependencies": {
