+++ frontend/src/index.css
@@ -1,1 +1,5 @@
 @import "tailwindcss";
+
+body {
+    @apply bg-gradient-to-tr from-blue-50 via-white to-green-100  text-gray-900 min-h-screen;
+}
